add type for Header and Profile
