+++ src/compo/Profile/Posts/MyPostsContainer.jsx
@@ -1,23 +1,23 @@
-import {addPostActionCreater} from '../../../redux/reducers/profileReducer'
-import MyPosts from './MyPosts'
-import {useDispatch, useSelector} from 'react-redux'
-import {postsDataState} from '../../../redux/selectors/profileSelector'
-
-
-const MyPostContainer = () => {
-
-    // STATE
-    const postsData = useSelector(state => postsDataState(state))
-
-    // DISPATCH
-    const dispatch = useDispatch()
-    const addPostActionCreater_ = (newPostText) => dispatch(addPostActionCreater(newPostText))
-
-    return <MyPosts postsData={postsData}  addPostAC={addPostActionCreater_}/>
-}
-
-
-
-export default MyPostContainer
-
-
+// import {addPostActionCreater} from '../../../redux/reducers/profileReducer'
+// import MyPosts from './MyPosts'
+// import {useDispatch, useSelector} from 'react-redux'
+// import {postsDataState} from '../../../redux/selectors/profileSelector'
+//
+//
+// const MyPostContainer = () => {
+//
+//     // STATE
+//     const postsData = useSelector(state => postsDataState(state))
+//
+//     // DISPATCH
+//     const dispatch = useDispatch()
+//     const addPostActionCreater_ = (newPostText) => dispatch(addPostActionCreater(newPostText))
+//
+//     return <MyPosts postsData={postsData}  addPostAC={addPostActionCreater_}/>
+// }
+//
+//
+//
+// export default MyPostContainer
+//
+//
