--- src/compo/Profile/Posts/PhotoPost/PhotoPost.jsx
@@ -1,9 +1,0 @@
-import postImg from './postImg.jpg'
-
-const PhotoPost = () => {
-    return (
-        <img src={ postImg } alt={"img_post"}/>
-    )
-}
-
-export default PhotoPost
