+++ src/compo/Profile/Posts/PhotoPost/PhotoPost.tsx
@@ -1,0 +1,9 @@
+import postImg from './postImg.jpg'
+
+const PhotoPost = () => {
+    return (
+        <img src={ postImg } alt={"img_post"}/>
+    )
+}
+
+export default PhotoPost
